Added appsettings support to web project and added ability to select API base url using the appsettings. Moved Installers to respective directories in web projects.
