+++ src/CookBook/CookBook.Web.App/Program.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Net.Http;
 using System.Threading.Tasks;
-using CookBook.Web.BL;
+using CookBook.Web.BL.Installers;
 using CookBook.Web.DAL;
+using CookBook.Web.DAL.Installers;
 using Microsoft.AspNetCore.Components.WebAssembly.Hosting;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 
 namespace CookBook.Web.App
@@ -20,8 +22,10 @@ namespace CookBook.Web.App
 
         private static void Install(WebAssemblyHostBuilder builder)
         {
+            var apiBaseUrl = builder.Configuration.GetValue<string>("ApiBaseUrl");
+
             new WebDALInstaller().Install(builder.Services);
-            new WebBLInstaller().Install(builder.Services);
+            new WebBLInstaller().Install(builder.Services, apiBaseUrl);
 
             builder.Services.AddScoped(_ => new HttpClient { BaseAddress = new Uri(builder.HostEnvironment.BaseAddress) });
             builder.Services.AddAutoMapper(typeof(WebBLInstaller));
