+++ src/CookBook/CookBook.Web.App/wwwroot/appsettings.json
@@ -1,0 +1,3 @@
+{
+    "ApiBaseUrl": ""
+}
