+++ src/CookBook/CookBook.Web.BL/ApiClient.cs
@@ -1,0 +1,13 @@
+﻿using System.Net.Http;
+
+namespace CookBook.Web.BL
+{
+    public partial class ApiClient
+    {
+        public ApiClient(HttpClient httpClient, string baseUrl)
+            : this(httpClient)
+        {
+            BaseUrl = baseUrl;
+        }
+    }
+}
