+++ src/CookBook/CookBook.Web.BL/Facades/IngredientFacade.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using AutoMapper;
-using CookBook.ApiClients;
 using CookBook.Common.Models;
 using CookBook.Web.DAL.Repositories;
