+++ src/CookBook/CookBook.Web.BL/GeneratedApiClient.cs
@@ -14,7 +14,7 @@ using CookBook.Common.Models;
 #pragma warning disable 8073 // Disable "CS8073 The result of the expression is always 'false' since a value of type 'T' is never equal to 'null' of type 'T?'"
 #pragma warning disable 3016 // Disable "CS3016 Arrays as attribute arguments is not CLS-compliant"
 
-namespace CookBook.ApiClients
+namespace CookBook.Web.BL
 {
     using System = global::System;
