+++ src/CookBook/CookBook.Web.BL/Installers/WebBLInstaller.cs
@@ -1,16 +1,16 @@
-﻿using CookBook.ApiClients;
+﻿using System.Net.Http;
 using CookBook.Common.BL.Facades;
 using CookBook.Common.Installers;
 using CookBook.Web.DAL;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace CookBook.Web.BL
+namespace CookBook.Web.BL.Installers
 {
-    public class WebBLInstaller : IInstaller
+    public class WebBLInstaller
     {
-        public void Install(IServiceCollection serviceCollection)
+        public void Install(IServiceCollection serviceCollection, string apiBaseUrl)
         {
-            serviceCollection.AddTransient<IApiClient, ApiClient>();
+            serviceCollection.AddTransient<IApiClient>(provider => new ApiClient(provider.GetService<HttpClient>(), apiBaseUrl));
             serviceCollection.AddSingleton<LocalDb>();
 
             serviceCollection.Scan(selector =>
