+++ src/CookBook/CookBook.Web.DAL/CookBook.Web.DAL.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.JSInterop" Version="6.0.0-preview.5.21301.17" />
+    <PackageReference Include="Microsoft.JSInterop" Version="6.0.0-preview.7.21378.6" />
   </ItemGroup>
 
   <ItemGroup>
